Cleanup `sparsity.py` and move tracing-based sparsity detections to `sparsity_tracing.py`
